% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_variation.R
\name{coefficient_of_variation}
\alias{coefficient_of_variation}
\title{Coefficient of variation calculator function}
\usage{
coefficient_of_variation(data)
}
\arguments{
\item{data}{Is a list/vector of only numeric values}
}
\value{
The calculated coefficient of variation
}
\description{
This is a function that calculates the coefficient of variation (sd/mean * 100) of a list of data
}
\details{
calculations are self-explanatory (see comments below), but we must ensure that the parameters are appropriate otherwise an error will be thrown
}
\examples{
coefficient_of_variation(c(1:100))

}
