% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr.R
\name{iqr}
\alias{iqr}
\title{IQR calculator function}
\usage{
iqr(data)
}
\arguments{
\item{data}{Is a list/vector of only numeric values}
}
\value{
The IQR of the data which is a numerical value
}
\description{
This is a function finds the IQR of the data
}
\details{
We call the "percentiles" function within this package to find the iqr = 3rd quantile - 1st quantile
}
\examples{
iqr(c(4,5,3,7,6,4,33,2,3,4))

}
