% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Mode finder function}
\usage{
mode(data)
}
\arguments{
\item{data}{Is a list/vector of only numeric values}
}
\value{
The calculated moment in the form of a numeric value
}
\description{
This is a function that finds the mode (the most frequent observation(s)) of a list of data
}
\details{
calculations are self-explanatory (see comments below), but we must ensure that the parameters are appropriate otherwise an error will be thrown
}
\examples{
mode(c(1,2,3,3,1,3,44,6))

}
