% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{Moment calculator function}
\usage{
moments(data, moment = 1)
}
\arguments{
\item{data}{Is a list/vector of only numeric values}

\item{moment}{Is the given moment that will be calculated (1 = first moment, 2= second moment, etc.) and is 1 if no parameter is passed}
}
\value{
The calculated moment in the form of a numeric value
}
\description{
This is a function that computes a given moment based off user input (first moment by default)
}
\details{
calculations are self-explanatory (see comments below), but we must ensure that the parameters are appropriate otherwise an error will be thrown
}
\examples{
moments(c(2,3,4,5,1,2,3,44,5,55,100,100))
moments(c(2,3,4,5,1,2,3,44,5,55,100,100), 4)

}
