% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentiles.R
\name{percentiles}
\alias{percentiles}
\title{Percentile finder function}
\usage{
percentiles(data, percentile = 0.5)
}
\arguments{
\item{data}{Is a list/vector of only numeric values}

\item{percentile}{Is the percentile in the data that we want to find and is 0.5 (median) if no parameter is passed}
}
\value{
The numerical value of the percentile
}
\description{
This is a function find a given percentile based off the users inputted data
}
\details{
Is the percentile in the data that we want to find and is 0.5 (median) if no parameter is passed
}
\examples{
percentiles(c(2:200))
percentiles(c(2:200), 0.9)

}
